+++ src/App/Osu/DiffCalc/AsyncPPCalculator.cpp
@@ -39,7 +39,8 @@ struct info_cache {
     bool ap{};
 
     // Results
-    std::unique_ptr<std::vector<DifficultyCalculator::DiffObject>> cachedDiffObjects{std::make_unique<std::vector<DifficultyCalculator::DiffObject>>()};
+    std::unique_ptr<std::vector<DifficultyCalculator::DiffObject>> cachedDiffObjects{
+        std::make_unique<std::vector<DifficultyCalculator::DiffObject>>()};
     pp_res info{};
     DifficultyCalculator::DifficultyAttributes diffattrs{};
 
@@ -232,7 +233,7 @@ void run_thread(const Sync::stop_token& stoken) {
                 .numCircles = map_for_rqt->iNumCircles,
                 .numSliders = map_for_rqt->iNumSliders,
                 .numSpinners = map_for_rqt->iNumSpinners,
-                .maxPossibleCombo = computed_ho->diffres.getTotalMaxCombo(),
+                .maxPossibleCombo = (i32)computed_ho->diffres.getTotalMaxCombo(),
                 .combo = rqt.comboMax,
                 .misses = rqt.numMisses,
                 .c300 = rqt.num300s,
